+++ android/src/main/java/com/flutter_webview_plugin/ObservableWebView.java
@@ -1,0 +1,49 @@
+package com.flutter_webview_plugin;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.webkit.WebView;
+
+public class ObservableWebView extends WebView {
+    private OnScrollChangedCallback mOnScrollChangedCallback;
+
+    public ObservableWebView(final Context context)
+    {
+        super(context);
+    }
+
+    public ObservableWebView(final Context context, final AttributeSet attrs)
+    {
+        super(context, attrs);
+    }
+
+    public ObservableWebView(final Context context, final AttributeSet attrs, final int defStyle)
+    {
+        super(context, attrs, defStyle);
+    }
+
+    @Override
+    protected void onScrollChanged(final int l, final int t, final int oldl, final int oldt)
+    {
+        super.onScrollChanged(l, t, oldl, oldt);
+        if(mOnScrollChangedCallback != null) mOnScrollChangedCallback.onScroll(l, t, oldl, oldt);
+    }
+
+    public OnScrollChangedCallback getOnScrollChangedCallback()
+    {
+        return mOnScrollChangedCallback;
+    }
+
+    public void setOnScrollChangedCallback(final OnScrollChangedCallback onScrollChangedCallback)
+    {
+        mOnScrollChangedCallback = onScrollChangedCallback;
+    }
+
+    /**
+     * Impliment in the activity/fragment/view that you want to listen to the webview
+     */
+    public static interface OnScrollChangedCallback
+    {
+        public void onScroll(int l, int t, int oldl, int oldt);
+    }
+}